fix church page routing bug
